Resolve Callback Race in ColdBootVisit
With the change implemented in #45, the WKNavigationDelegate on
ColdBootVisit is removed when the WebViewPageLoadDelegate method
webView:didLoadPageWithRestorationIdentifier: is called. This
callback originates from the WKScriptMessageHandler inside
Turbolinks.WebView.

In production, I'm seeing cases where this
method can be called *before* the webView:didFinish: callback
of WKNavigationDelegate. As a result, the SessionDelegate
sessionDidFinishRequest: method is never called. The behavior
is indeterminate. It appears to be a race condition with no
guarantee of which will execute first.

Ensuring the delegate is not removed until after the webview finishes
loading prevents this callback from getting dropped.